+++ src/index.js
@@ -1,5 +1,7 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
+import { CookiesProvider } from 'react-cookie'
+import { GoogleOAuthProvider } from '@react-oauth/google'
 import './index.css'
 import App from './App'
 // eslint-disable-next-line import/no-named-as-default, import/no-named-as-default-member
@@ -7,9 +9,13 @@ import reportWebVitals from './reportWebVitals'
 
 ReactDOM.render(
   // eslint-disable-next-line react/jsx-filename-extension
-  <React.StrictMode>
-    <App />
-  </React.StrictMode>,
+  <CookiesProvider defaultSetOptions={{ path: '/' }}>
+    <GoogleOAuthProvider clientId={process.env.REACT_APP_GOOGLE_CLIENT_ID}>
+      <React.StrictMode>
+        <App />
+      </React.StrictMode>
+    </GoogleOAuthProvider>
+  </CookiesProvider>,
   // eslint-disable-next-line no-undef
   document.getElementById('root'),
 )