+++ src/views/Renamer.jsx
@@ -1,4 +1,5 @@
 import React, { useState } from 'react'
+import { useCookies } from 'react-cookie'
 import { Badge, Card, Form } from 'react-bootstrap'
 import Button from 'react-bootstrap/Button'
 import PropTypes from 'prop-types'
@@ -18,6 +19,8 @@ import TwilioConfiguration from '../components/Renamer/TwilioConfiguration'
 
 export default function Renamer(props) {
   const { activatedDevices, particleToken, clickupToken, clickupListID, environment, modifyActivatedDevice } = props
+
+  const [cookies] = useCookies(['googleIdToken'])
 
   const blankActivatedDevice = ActivatedDevice.BlankDevice()
 
@@ -169,7 +172,7 @@ export default function Renamer(props) {
     if (twilioCheck) {
       setTwilioStatus('waiting')
       // purchase twilio number
-      const twilioResponse = await purchaseSensorTwilioNumberByAreaCode(twilioAreaCode, locationID, environment, clickupToken)
+      const twilioResponse = await purchaseSensorTwilioNumberByAreaCode(twilioAreaCode, locationID, environment, cookies.googleIdToken)
       if (twilioResponse.message === 'success') {
         setTwilioStatus(twilioResponse.phoneNumber)
         newTwilioPhoneNumber = twilioResponse.phoneNumber
@@ -195,7 +198,7 @@ export default function Renamer(props) {
 
       newTwilioPhoneNumber = twilioCheck ? newTwilioPhoneNumber : twilioPhoneNumber
       const databaseInsert = await insertSensorLocation(
-        clickupToken,
+        cookies.googleIdToken,
         password,
         locationID,
         displayName,
@@ -334,7 +337,6 @@ export default function Renamer(props) {
               dashboardCheck={dashboardCheck}
               client={client}
               changeClient={changeClient}
-              clickupToken={clickupToken}
               displayName={displayName}
               changeDisplayName={changeDisplayName}
               password={password}